core/sxmo_statusbar: handle USR1 as early as possible

+++ scripts/core/sxmo_statusbar.sh
@@ -1,15 +1,17 @@
 #!/bin/sh
 
+forceupdate() {
+	[ -z "$SLEEPID" ] || kill "$SLEEPID"
+}
+# trap USR1 as early as possible in case something sends the signal early on
+# startup
+trap "forceupdate" USR1
+
 # include common definitions
 # shellcheck source=scripts/core/sxmo_common.sh
 . "$(dirname "$0")/sxmo_common.sh"
 
 POLL_TIME="${SXMO_STATUSBAR_POLL_TIME:-10}"
-
-forceupdate() {
-	kill "$SLEEPID"
-}
-trap "forceupdate" USR1
 
 setbar() {
 	case "$WM" in
